Fix round calculation to use floor instead of ceiling
When there are three cards before the next epidemic or escalation card,
it implies that it will come at the end of the next round. Thus, the UI
should display that there is only one round before it happens instead of
two as there is only one round of relief instead of two.

Signed-off-by: Remy Suen <remy.suen@gmail.com>